refactor: rename MovieDetailsRemoteImpl package
