+++ app/src/main/java/com/london/app/di/DataSourceModule.kt
@@ -6,7 +6,7 @@ import com.london.data.BuildConfig
 import com.london.data.datasource.remote.details.actordetails.ActorDetailsRemoteDataSource
 import com.london.data.datasource.remote.details.actordetails.ActorDetailsRemoteDataSourceImpl
 import com.london.data.datasource.remote.details.moviedetails.MovieDetailsRemote
-import com.london.data.datasource.remote.details.moviedetails.model.MovieDetailsRemoteImpl
+import com.london.data.datasource.remote.details.moviedetails.MovieDetailsRemoteImpl
 import com.london.data.datasource.remote.details.tvshowdetails.TvShowDetailsRemoteDataSource
 import com.london.data.datasource.remote.details.tvshowdetails.TvShowDetailsRemoteDataSourceImpl
 import com.london.data.datasource.remote.search.SearchRemoteDataSource
