+++ data/src/test/java/com/london/data/datasource/remote/moviedetails/MovieDetailsResponseImplTest.kt
@@ -2,7 +2,7 @@ package com.london.data.datasource.remote.moviedetails
 
 import android.util.Log
 import com.google.common.truth.Truth.assertThat
-import com.london.data.datasource.remote.details.moviedetails.model.MovieDetailsRemoteImpl
+import com.london.data.datasource.remote.details.moviedetails.MovieDetailsRemoteImpl
 import com.london.data.datasource.remote.details.moviedetails.model.moviecast.MovieCastResponse
 import com.london.data.datasource.remote.details.moviedetails.model.moviecast.MovieImages
 import com.london.data.datasource.remote.details.moviedetails.model.moviedetails.MovieDetailsResponse
